[refactor] 게시글 폼에서 공통적으로 적용되는 이벤트를 각 작성/수정 페이지의 js가 아니라 post 컴포넌트에서 등록하도록 리팩토링
